Polished the moderator list
